Cleanup UI into separate files
