+++ NeuralAmpModeler/config.h
@@ -24,10 +24,14 @@
 #define PLUG_DOES_STATE_CHUNKS 0
 #define PLUG_HAS_UI 1
 #define PLUG_WIDTH 600
-#define PLUG_HEIGHT 600
+#define PLUG_HEIGHT 400
 #define PLUG_FPS 60
 #define PLUG_SHARED_RESOURCES 0
 #define PLUG_HOST_RESIZE 0
+#define PLUG_MIN_WIDTH (PLUG_WIDTH/3)
+#define PLUG_MIN_HEIGHT (PLUG_HEIGHT/3)
+#define PLUG_MAX_WIDTH (PLUG_WIDTH*3)
+#define PLUG_MAX_HEIGHT (PLUG_HEIGHT*3)
 
 #define AUV2_ENTRY NeuralAmpModeler_Entry
 #define AUV2_ENTRY_STR "NeuralAmpModeler_Entry"
